Unify utilities into a single service for containerization (#372)
* WIP - Adding back migrate check, working to unify services into one image

* Unified utilities into one service

* Renamed to api-service

* Renaming to KEEP_ALIVE

* Removing obsolete dockerfile

* Pass full opts object to migrate

* Renaming to service.dockerfile

* Adding capabilities for running full system with docker compose

* Invoking all services to run as full

* Completed full system lifecycle, updated README

* Formatting tweak to README

* Reordered sections in README

* Only trigger ci/cd tests on relevant changes

* Fix on: syntax in test action

+++ src/migrate/index.js
@@ -1,19 +1,53 @@
+const objectUtils = require('lodash/object')
 const Asset = require('../models/Asset')
 const Market = require('../models/Market')
 const assets = require('./data/assets.json')
 const markets = require('./data/markets.json')
 
+const defaultLog = true
+const defaultVerbose = false
+const defaultForce = false
 const logHeader = '[MIGRATE]'
 
 module.exports.run = async (options = {}) => {
-  console.log(`${logHeader} Seeding data...`)
-  await Asset.deleteMany({})
-  const newAssets = await Asset.insertMany(assets, { ordered: false })
-  console.log(`${logHeader} ${newAssets.length} assets have been set`)
+  console.log(`${logHeader} Running migrate with options:`, options)
+  const log = objectUtils.get(options, 'log', defaultLog)
+  const verbose = objectUtils.get(options, 'verbose', defaultVerbose)
+  const force = objectUtils.get(options, 'force', defaultForce)
 
-  await Market.deleteMany({})
-  const newMarkets = await Market.insertMany(markets, { ordered: false })
-  console.log(`${logHeader} ${newMarkets.length} markets have been set`)
+  const keepAlive = process.env.KEEP_ALIVE || false
+  console.log(`${logHeader} keep alive?`, keepAlive)
 
-  process.exit()
+  if (!force && await hasData({ verbose })) {
+    if (log) console.log(`${logHeader} Data is already seeded.`)
+  } else {
+    if (log) console.log(`${logHeader} Seeding data...`)
+    await Asset.deleteMany({})
+    const newAssets = await Asset.insertMany(assets, { ordered: false })
+    if (log) console.log(`${logHeader} ${newAssets.length} assets have been set`)
+
+    await Market.deleteMany({})
+    const newMarkets = await Market.insertMany(markets, { ordered: false })
+    if (log) console.log(`${logHeader} ${newMarkets.length} markets have been set`)
+  }
+
+  if (!keepAlive) process.exit()
+}
+
+async function hasData (options = {}) {
+  const verbose = options.verbose
+  let result = true
+
+  if (verbose) console.log(`${logHeader} Checking for existing assets`)
+
+  const existingAssets = await Asset.find({}).exec()
+
+  if (!existingAssets || existingAssets.length === 0) {
+    if (verbose) console.log('No assets found.')
+    result = false
+  } else {
+    if (verbose) console.log('Existing assets:', existingAssets)
+  }
+
+  return result
 }
